chore: update file structure + remove unused files
