+++ src/components/analytics/AirPollutionChart.tsx
@@ -10,22 +10,25 @@ import {
   YAxis,
 } from "recharts";
 
-import type { TimeSeries, Metric } from "@/components/charts/types";
-import { getTickFormat } from "@/lib/timeSeries";
+import type { TimeSeries, Metric, TimeSeriesDataKey } from "@/components/analytics/types";
+import { getTickFormat } from "@/lib/utils/timeSeries";
 
-interface EnvironmentalConditionsChartProps {
+interface AirPollutionChartProps<T extends TimeSeriesDataKey> {
   timeSeries: TimeSeries;
-  metrics: Metric[];
+  metrics: Metric<T>[];
 }
 
-function EnvironmentalConditionsChart({ timeSeries, metrics }: EnvironmentalConditionsChartProps) {
+function AirPollutionChart<T extends TimeSeriesDataKey>({
+  timeSeries,
+  metrics,
+}: AirPollutionChartProps<T>) {
   const { data, isError, isLoading, error } = timeSeries;
   const activeMetrics = metrics.filter((metric) => metric.enabled);
 
   if (isLoading) {
     return (
       <div className="flex h-75 items-center justify-center">
-        <span className="text-sm text-muted-foreground">Loading environmental data...</span>
+        <span className="text-sm text-muted-foreground">Loading air quality data...</span>
       </div>
     );
   }
@@ -59,7 +62,7 @@ function EnvironmentalConditionsChart({ timeSeries, metrics }: EnvironmentalCond
     data.start_time,
     data.interval_unit,
     data.interval_length,
-    data.temperature.length,
+    data.pm1p0.length,
   );
   const formatTick = (value: number) =>
     tickFormat ? format(new Date(value), tickFormat) : new Date(value).toLocaleString();
@@ -67,11 +70,15 @@ function EnvironmentalConditionsChart({ timeSeries, metrics }: EnvironmentalCond
   return (
     <ResponsiveContainer width="100%" height={280}>
       <LineChart
-        data={data.temperature.map((temperature, i) => ({
-          time: addInterval(data.start_time, i * data.interval_length).getTime(),
-          temperature,
-          humidity: data.humidity[i],
-        }))}
+        data={data.pm1p0.map((pm1p0, i) => {
+          return {
+            time: addInterval(data.start_time, i * data.interval_length).getTime(),
+            pm1p0,
+            pm2p5: data.pm2p5[i],
+            pm4p0: data.pm4p0[i],
+            pm10: data.pm10[i],
+          };
+        })}
       >
         <CartesianGrid strokeDasharray="3 3" stroke="rgba(255,255,255,0.06)" />
         <XAxis
@@ -86,28 +93,14 @@ function EnvironmentalConditionsChart({ timeSeries, metrics }: EnvironmentalCond
           tickLine={false}
         />
         <YAxis
-          yAxisId="left"
+          domain={[0, "auto"]}
           tick={{ fill: "var(--color-muted-foreground)", fontSize: 12 }}
           axisLine={false}
           tickLine={false}
           label={{
-            value: "Temperature (°C)",
+            value: "μg/m³",
             angle: -90,
             position: "insideLeft",
-            style: { fontSize: 12, fill: "var(--color-muted-foreground)" },
-          }}
-        />
-        <YAxis
-          yAxisId="right"
-          orientation="right"
-          domain={[0, 100]}
-          tick={{ fill: "var(--color-muted-foreground)", fontSize: 12 }}
-          axisLine={false}
-          tickLine={false}
-          label={{
-            value: "Humidity (%)",
-            angle: 90,
-            position: "insideRight",
             style: { fontSize: 12, fill: "var(--color-muted-foreground)" },
           }}
         />
@@ -125,7 +118,6 @@ function EnvironmentalConditionsChart({ timeSeries, metrics }: EnvironmentalCond
         {activeMetrics.map((metric) => (
           <Line
             key={metric.key}
-            yAxisId={metric.key === "temperature" ? "left" : "right"}
             type="monotone"
             dataKey={metric.key}
             stroke={metric.color}
@@ -139,4 +131,4 @@ function EnvironmentalConditionsChart({ timeSeries, metrics }: EnvironmentalCond
   );
 }
 
-export { EnvironmentalConditionsChart };
+export { AirPollutionChart };
