+++ src/index.css
@@ -21,11 +21,6 @@
   --color-success: oklch(56% 0.1218 155.86);
   --color-warn: oklch(74.74% 0.1796 57.14);
   --color-destructive: oklch(57.85% 0.2138 27.17);
-  --color-chart-1: oklch(0.871 0.15 154.449);
-  --color-chart-2: oklch(92.271% 0.12602 156.575);
-  --color-chart-3: oklch(0.627 0.194 149.214);
-  --color-chart-4: oklch(0.527 0.154 150.069);
-  --color-chart-5: oklch(0.448 0.119 151.328);
 }
 
 @layer base {
