+++ src/lib/hooks/useEnvironmentTimeSeries.ts
@@ -1,12 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
 
 import { env } from "@/env";
-import { getHeaders } from "@/lib/headers";
-import { addQueryParameters } from "@/lib/queryParameters";
 import type {
   EnvironmentTimeSeriesResponse,
   GetEnvironmentTimeSeriesParameters,
 } from "@/lib/types/api";
+import { getHeaders } from "@/lib/utils/headers";
+import { addQueryParameters } from "@/lib/utils/queryParameters";
 
 function useEnvironmentTimeSeries(queryParams: GetEnvironmentTimeSeriesParameters) {
   return useQuery({
