+++ src/lib/hooks/useObservationsTimeSeries.tsx
@@ -1,12 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
 
 import { env } from "@/env";
-import { getHeaders } from "@/lib/headers";
-import { addQueryParameters } from "@/lib/queryParameters";
 import type {
   GetObservationsTimeSeriesParameters,
   ObservationsTimeSeriesResponse,
 } from "@/lib/types/api";
+import { getHeaders } from "@/lib/utils/headers";
+import { addQueryParameters } from "@/lib/utils/queryParameters";
 
 function useObservationsTimeSeries(queryParameters: GetObservationsTimeSeriesParameters) {
   return useQuery({
