+++ src/lib/types/api.ts
@@ -199,12 +199,6 @@ interface GetObservationCountParameters {
   selected_taxa?: string[];
 }
 
-// Should be removed - migrate to useFilters()
-
-type DatePreset = "all" | "24h" | "7d" | "30d" | "custom";
-
-type WorkspaceTab = "overview" | "analytics" | "observations";
-
 export type {
   Location,
   Environment,
@@ -229,6 +223,4 @@ export type {
   GetObservationsParameters,
   GetObservationCountParameters,
   TaxonomyLevel,
-  DatePreset,
-  WorkspaceTab,
 };
