+++ src/lib/utils/heatmap.ts
